make sure tmp is made

+++ backend/routes/lynxchan_minimal.php
@@ -103,8 +103,11 @@ $router->post('/createBoard', function($request) {
 });
 
 $router->post('/files', function($request) {
+  // make sure tmp is made
+  if (!file_exists('storage/tmp')) {
+    return sendResponse(array(), 400, 'Backend server is not ready for files');
+  }
   $hash = hash_file('sha256', $_FILES['files']['tmp_name']);
-  // FIXME: make sure tmp is made
   move_uploaded_file($_FILES['files']['tmp_name'], 'storage/tmp/'.$hash);
   $data=array(
     'type' => $_FILES['files']['type'],
